+++ Tests/Unit/Middleware/TrackingCodeInjectorTest.php
@@ -11,28 +11,31 @@ declare(strict_types=1);
 
 namespace Brotkrueml\MatomoIntegration\Tests\Unit\Hooks\PageRenderer;
 
-use Brotkrueml\MatomoIntegration\Adapter\ApplicationType;
 use Brotkrueml\MatomoIntegration\Code\JavaScriptTrackingCodeBuilder;
 use Brotkrueml\MatomoIntegration\Code\NoScriptTrackingCodeBuilder;
 use Brotkrueml\MatomoIntegration\Code\TagManagerCodeBuilder;
-use Brotkrueml\MatomoIntegration\Hooks\PageRenderer\TrackingCodeInjector;
-use PHPUnit\Framework\MockObject\MockObject;
+use Brotkrueml\MatomoIntegration\Middleware\TrackingCodeInjection;
 use PHPUnit\Framework\MockObject\Stub;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
-use TYPO3\CMS\Core\Page\PageRenderer;
+use Psr\Http\Server\RequestHandlerInterface;
+use TYPO3\CMS\Core\Http\Response;
 use TYPO3\CMS\Core\Site\Entity\Site;
 
 final class TrackingCodeInjectorTest extends TestCase
 {
-    /** @var Stub|ApplicationType */
-    private $applicationTypeStub;
+    private const ORIGINAL_RESPONSE_CONTENTS = '<html lang="en"><head><title>Some title</title></head><body>Some body</body></html>';
 
     /** @var Stub|Site */
     private $siteStub;
 
     /** @var Stub|ServerRequestInterface */
     private $requestStub;
+
+    /** @var Stub|RequestHandlerInterface */
+    private $handlerStub;
+
+    private Response $response;
 
     /** @var Stub|JavaScriptTrackingCodeBuilder */
     private $javaScriptTrackingCodeBuilderStub;
@@ -43,16 +46,8 @@ final class TrackingCodeInjectorTest extends TestCase
     /** @var Stub|TagManagerCodeBuilder */
     private $tagManagerCodeBuilderStub;
 
-    /** @var MockObject|PageRenderer */
-    private $pageRendererMock;
-
     protected function setUp(): void
     {
-        $this->applicationTypeStub = $this->createStub(ApplicationType::class);
-        $this->applicationTypeStub
-            ->method('isBackend')
-            ->willReturn(false);
-
         $this->siteStub = $this->createStub(Site::class);
 
         $this->requestStub = $this->createStub(ServerRequestInterface::class);
@@ -60,6 +55,16 @@ final class TrackingCodeInjectorTest extends TestCase
             ->method('getAttribute')
             ->with('site')
             ->willReturn($this->siteStub);
+
+        $this->response = new Response();
+        $this->response = $this->response->withHeader('Content-Type', 'text/html; charset=utf-8');
+        $this->response->getBody()->write(self::ORIGINAL_RESPONSE_CONTENTS);
+
+        $this->handlerStub = $this->createStub(RequestHandlerInterface::class);
+        $this->handlerStub
+            ->method('handle')
+            ->with($this->requestStub)
+            ->willReturn($this->response);
 
         $this->javaScriptTrackingCodeBuilderStub = $this->createStub(JavaScriptTrackingCodeBuilder::class);
         $this->javaScriptTrackingCodeBuilderStub
@@ -75,50 +80,33 @@ final class TrackingCodeInjectorTest extends TestCase
         $this->tagManagerCodeBuilderStub
             ->method('setConfiguration')
             ->willReturn($this->tagManagerCodeBuilderStub);
-
-        $this->pageRendererMock = $this->createMock(PageRenderer::class);
     }
 
     /**
      * @test
      */
-    public function executeDoesNothingWhenInBackend(): void
+    public function processReturnsOriginalResponseContentsWhenContentTypeIsNotTextHtml(): void
     {
-        $applicationTypeStub = $this->createStub(ApplicationType::class);
-        $applicationTypeStub
-            ->method('isBackend')
-            ->willReturn(true);
+        $this->response = $this->response->withHeader('Content-Type', 'application/json');
 
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addHeaderData');
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addFooterData');
-
-        $subject = new TrackingCodeInjector(
-            $applicationTypeStub,
-            $this->requestStub,
+        $subject = new TrackingCodeInjection(
             $this->javaScriptTrackingCodeBuilderStub,
             $this->noScriptTrackingCodeBuilderStub,
             $this->tagManagerCodeBuilderStub
         );
-        $params = [];
-        $subject->execute($params, $this->pageRendererMock);
+
+        $actual = $subject->process($this->requestStub, $this->handlerStub);
+        $body = $actual->getBody();
+        $body->rewind();
+
+        self::assertSame(self::ORIGINAL_RESPONSE_CONTENTS, $body->getContents());
     }
 
     /**
      * @test
      */
-    public function executeDoesNothingWhenConfigurationHasNoValidUrl(): void
+    public function processReturnsOriginalResponseContentsWhenConfigurationHasNoValidUrl(): void
     {
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addHeaderData');
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addFooterData');
-
         $configuration = [
             'matomoIntegrationUrl' => 'some invalid url',
             'matomoIntegrationSiteId' => 42,
@@ -128,29 +116,24 @@ final class TrackingCodeInjectorTest extends TestCase
             ->method('getConfiguration')
             ->willReturn($configuration);
 
-        $subject = new TrackingCodeInjector(
-            $this->applicationTypeStub,
-            $this->requestStub,
+        $subject = new TrackingCodeInjection(
             $this->javaScriptTrackingCodeBuilderStub,
             $this->noScriptTrackingCodeBuilderStub,
             $this->tagManagerCodeBuilderStub
         );
-        $params = [];
-        $subject->execute($params, $this->pageRendererMock);
+
+        $actual = $subject->process($this->requestStub, $this->handlerStub);
+        $body = $actual->getBody();
+        $body->rewind();
+
+        self::assertSame(self::ORIGINAL_RESPONSE_CONTENTS, $body->getContents());
     }
 
     /**
      * @test
      */
-    public function executeDoesNothingWhenConfigurationHasNoValidSiteId(): void
+    public function processReturnsOriginalResponseContentsWhenConfigurationHasNoValidSiteId(): void
     {
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addHeaderData');
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addFooterData');
-
         $configuration = [
             'matomoIntegrationUrl' => 'https://example.org/',
             'matomoIntegrationSiteId' => 0,
@@ -160,30 +143,24 @@ final class TrackingCodeInjectorTest extends TestCase
             ->method('getConfiguration')
             ->willReturn($configuration);
 
-        $subject = new TrackingCodeInjector(
-            $this->applicationTypeStub,
-            $this->requestStub,
+        $subject = new TrackingCodeInjection(
             $this->javaScriptTrackingCodeBuilderStub,
             $this->noScriptTrackingCodeBuilderStub,
             $this->tagManagerCodeBuilderStub
         );
-        $params = [];
-        $subject->execute($params, $this->pageRendererMock);
+
+        $actual = $subject->process($this->requestStub, $this->handlerStub);
+        $body = $actual->getBody();
+        $body->rewind();
+
+        self::assertSame(self::ORIGINAL_RESPONSE_CONTENTS, $body->getContents());
     }
 
     /**
      * @test
      */
-    public function executeAddsJavaScriptTrackingCodeWithoutTagManagerToHeaderDataCorrectly(): void
+    public function processInjectsJavaScriptTrackingCodeWithoutTagManagerIntoHeadCorrectly(): void
     {
-        $this->pageRendererMock
-            ->expects(self::once())
-            ->method('addHeaderData')
-            ->with('<script>/* some tracking code */</script>');
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addFooterData');
-
         $configuration = [
             'matomoIntegrationUrl' => 'https://example.org/',
             'matomoIntegrationSiteId' => 42,
@@ -197,30 +174,28 @@ final class TrackingCodeInjectorTest extends TestCase
             ->method('getTrackingCode')
             ->willReturn('/* some tracking code */');
 
-        $subject = new TrackingCodeInjector(
-            $this->applicationTypeStub,
-            $this->requestStub,
+        $subject = new TrackingCodeInjection(
             $this->javaScriptTrackingCodeBuilderStub,
             $this->noScriptTrackingCodeBuilderStub,
             $this->tagManagerCodeBuilderStub
         );
-        $params = [];
-        $subject->execute($params, $this->pageRendererMock);
+
+        $actual = $subject->process($this->requestStub, $this->handlerStub);
+        $body = $actual->getBody();
+        $body->rewind();
+
+        self::assertSame(
+            '<html lang="en"><head><script>/* some tracking code */</script>
+<title>Some title</title></head><body>Some body</body></html>',
+            $body->getContents()
+        );
     }
 
     /**
      * @test
      */
-    public function executeAddsJavaScriptTrackingCodeWithTagManagerToHeaderDataCorrectly(): void
+    public function processIbjectsJavaScriptTrackingCodeWithTagManagerIntoHeadCorrectly(): void
     {
-        $this->pageRendererMock
-            ->expects(self::once())
-            ->method('addHeaderData')
-            ->with('<script>/* some tracking code *//* some tag manager code */</script>');
-        $this->pageRendererMock
-            ->expects(self::never())
-            ->method('addFooterData');
-
         $configuration = [
             'matomoIntegrationUrl' => 'https://example.org/',
             'matomoIntegrationSiteId' => 42,
@@ -239,30 +214,28 @@ final class TrackingCodeInjectorTest extends TestCase
             ->method('getCode')
             ->willReturn('/* some tag manager code */');
 
-        $subject = new TrackingCodeInjector(
-            $this->applicationTypeStub,
-            $this->requestStub,
+        $subject = new TrackingCodeInjection(
             $this->javaScriptTrackingCodeBuilderStub,
             $this->noScriptTrackingCodeBuilderStub,
             $this->tagManagerCodeBuilderStub
         );
-        $params = [];
-        $subject->execute($params, $this->pageRendererMock);
+
+        $actual = $subject->process($this->requestStub, $this->handlerStub);
+        $body = $actual->getBody();
+        $body->rewind();
+
+        self::assertSame(
+            '<html lang="en"><head><script>/* some tracking code *//* some tag manager code */</script>
+<title>Some title</title></head><body>Some body</body></html>',
+            $body->getContents()
+        );
     }
 
     /**
      * @test
      */
-    public function executeAddsNoScriptTrackingCodeToFooterDataCorrectly(): void
+    public function processInjectsNoScriptTrackingCodeIntoBodyCorrectly(): void
     {
-        $this->pageRendererMock
-            ->expects(self::once())
-            ->method('addHeaderData');
-        $this->pageRendererMock
-            ->expects(self::once())
-            ->method('addFooterData')
-            ->with('<noscript><!-- some tracking code --></noscript>');
-
         $configuration = [
             'matomoIntegrationUrl' => 'https://example.org/',
             'matomoIntegrationSiteId' => 42,
@@ -273,18 +246,29 @@ final class TrackingCodeInjectorTest extends TestCase
             ->method('getConfiguration')
             ->willReturn($configuration);
 
+        $this->javaScriptTrackingCodeBuilderStub
+            ->method('getTrackingCode')
+            ->willReturn('/* some tracking code */');
+
         $this->noScriptTrackingCodeBuilderStub
             ->method('getTrackingCode')
             ->willReturn('<!-- some tracking code -->');
 
-        $subject = new TrackingCodeInjector(
-            $this->applicationTypeStub,
-            $this->requestStub,
+        $subject = new TrackingCodeInjection(
             $this->javaScriptTrackingCodeBuilderStub,
             $this->noScriptTrackingCodeBuilderStub,
             $this->tagManagerCodeBuilderStub
         );
-        $params = [];
-        $subject->execute($params, $this->pageRendererMock);
+
+        $actual = $subject->process($this->requestStub, $this->handlerStub);
+        $body = $actual->getBody();
+        $body->rewind();
+
+        self::assertSame(
+            '<html lang="en"><head><script>/* some tracking code */</script>
+<title>Some title</title></head><body>Some body<noscript><!-- some tracking code --></noscript>
+</body></html>',
+            $body->getContents()
+        );
     }
 }