+++ Tests/Unit/EventListener/TrackJavaScriptErrorsTest.php
@@ -1,0 +1,63 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "matomo_integration" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\MatomoIntegration\Tests\Unit\EventListener;
+
+use Brotkrueml\MatomoIntegration\Entity\Configuration;
+use Brotkrueml\MatomoIntegration\Event\AfterTrackPageViewEvent;
+use Brotkrueml\MatomoIntegration\EventListener\TrackJavaScriptErrors;
+use PHPUnit\Framework\TestCase;
+
+final class TrackJavaScriptErrorsTest extends TestCase
+{
+    private TrackJavaScriptErrors $subject;
+
+    protected function setUp(): void
+    {
+        $this->subject = new TrackJavaScriptErrors();
+    }
+
+    /**
+     * @test
+     */
+    public function disabledOption(): void
+    {
+        $configuration = Configuration::createFromSiteConfiguration([
+            'matomoIntegrationUrl' => 'https://www.example.net/',
+            'matomoIntegrationSiteId' => 123,
+        ]);
+
+        $event = new AfterTrackPageViewEvent($configuration);
+        $this->subject->__invoke($event);
+
+        $actual = $event->getMatomoMethodCalls();
+        self::assertCount(0, $actual);
+    }
+
+    /**
+     * @test
+     */
+    public function enabledOption(): void
+    {
+        $configuration = Configuration::createFromSiteConfiguration([
+            'matomoIntegrationUrl' => 'https://www.example.net/',
+            'matomoIntegrationSiteId' => 123,
+            'matomoIntegrationOptions' => 'trackJavaScriptErrors',
+        ]);
+
+        $event = new AfterTrackPageViewEvent($configuration);
+        $this->subject->__invoke($event);
+
+        $actual = $event->getMatomoMethodCalls();
+        self::assertCount(1, $actual);
+        self::assertSame('_paq.push(["enableJSErrorTracking"]);', (string)$actual[0]);
+    }
+}