+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Use middleware instead of hook for tracking code injection
+
 ## [1.0.0] - 2021-08-30
 
 Initial release