+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1] - 2021-09-28
+
 ### Fixed
 - Remove suggests to avoid dependency issue with EXT:matomo_widgets v1.1.2
 
@@ -14,5 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/brotkrueml/typo3-matomo-integration/releases/tag/v1.0.0